% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCSBM.R
\name{DCSBM}
\alias{DCSBM}
\title{DCSBM}
\usage{
DCSBM(n, k = 2, P, sizes = c(round(n/2), n - round(n/2)),
  random.community.assignment = c(FALSE, TRUE), community.labels = NULL,
  delta = rep(0, k), edge.list = c(FALSE, TRUE))
}
\arguments{
\item{n:}{number of nodes in the graph}

\item{k:}{number of communities in the graph}

\item{P:}{the k x k matrix of probabilities whose i,jth entry specifies the probability of connection between nodes in community i and community j}

\item{sizes:}{a numeric vector of length k whose ith entry specifies the size of the ith community. The entries must add to n.}

\item{random.community.assignment:}{a logical that specifies whether or not community labels are determined at random. Default is FALSE.}

\item{community.labels:}{an integer vector of length n whose ith entry is the community label of the ith vertex. Default is NULL. If provided, community labels are no longer assigned.}

\item{delta:}{a numeric vector of length k whose values must be between 0 and 1. Theta parameters for community r are generated as an iid sample from a U(0 + delta, 1 - delta) distribution}

\item{edge.list:}{a logical that specifies whether or not the adjacency matrix should be returned as an edge list.}
}
\value{
a list containing the objects 
\itemize{
   \item Adjacency: the adjacency matrix (or edge list if edge.list == TRUE) of the generated network
   \item Thetas: an n x 1 vector with values of each theta
   \item Membership: an n x 1 vector specifying community membership of each node
}
}
\description{
Simulate an undirected graph realization from the degree corrected stochastic block random graph model. Edge weights are discrete valued and are generated independently where e_ij ~ Poisson(theta_i*theta_j*P_{c_i, c_j})
}
\examples{
net <- DCSBM(n = 500, k = 2, P = cbind(c(0.10, 0.01), c(0.02, 0.075)),
             sizes = c(200, 300), random.community.assignment = FALSE,
             delta = c(0.2, 0.7), edge.list = FALSE)
image(Matrix(net$Adjacency))
}
\author{
James D. Wilson and Nathaniel T. Stevens
}
\references{
\itemize{
    \item Wilson, James D., Stevens, Nathaniel T., and Woodall, William H. (2016). “Modeling and estimating change in temporal networks via a dynamic degree corrected stochastic block model.” 
    arXiv Preprint: http://arxiv.org/abs/1605.04049
}
}
\keyword{community}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{monitoring}
\keyword{network}
\keyword{random}

