% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_DCSBM.R
\name{MLE.DCSBM}
\alias{MLE.DCSBM}
\title{MLE.DCSBM}
\usage{
MLE.DCSBM(Adjacency.list, community.array, T, k)
}
\arguments{
\item{Adjacency.list:}{list of adjacency matrices in the observed dynamic network}

\item{community.array:}{an array of length T whose tth entry is a numeric vector of length n specifying community labels at time t}

\item{T:}{number of graphs in the temporal sequence}

\item{k:}{number of communities (fixed accross time) Note, this is the pre-conceived idea of how many communities there are.}
}
\value{
a list containing the objects 
\itemize{
   \item P.hat.array: an array of length T whose tth entry is the estimated MLE of P for the tth network
   \item delta.hat.array: an array of length T whose tth entry are the estimated MLEs of the delta parameters for the tth network
   \item delta.hat.global: a numeric of length T whose tth entry is the estimated MLE of the overall standard deviation of the theta parameters for the tth network
}
}
\description{
Estimate the maximum likelihood estimators for P and delta at each time point in the list of adjacency matrices
}
\examples{
#Generate a collection of 50 networks with a change at time 25. The change is a local 
#change in connection propensity in community 1
n <- 100
P.old <- cbind(c(0.10, 0.01), c(0.02, 0.075))
P.new <- cbind(c(0.20, 0.025), c(0.02, 0.075))
P.array <- array(c(replicate(25, P.old), replicate(25, P.new)), dim = c(2, 2, 50))
community.array <- array(rep(c(rep(1, 50), rep(2, 50)), 50), dim = c(1, 100, 50))
delta.array <- array(rep(rep(0.2, 2), 50), dim = c(1, 2, 50))

dynamic.net <- dynamic.DCSBM(n = 100, T = 50, P.array = P.array,
                             community.array = community.array,
                             delta.array = delta.array, edge.list = FALSE)
image(Matrix(dynamic.net$Adjacency.list[[1]]))
image(Matrix(dynamic.net$Adjacency.list[[30]]))

#Estimate the MLEs
MLEs.example <- MLE.DCSBM(dynamic.net$Adjacency.list, community.array = community.array,
                           T = 50, k = 2)
}
\author{
James D. Wilson and Nathaniel T. Stevens
}
\references{
\itemize{
    \item Wilson, James D., Stevens, Nathaniel T., and Woodall, William H. (2016). “Modeling and estimating change in temporal networks via a dynamic degree corrected stochastic block model.” 
    arXiv Preprint: http://arxiv.org/abs/1605.04049
}
}
\keyword{community}
\keyword{control}
\keyword{detection,}
\keyword{graph}
\keyword{model,}
\keyword{monitoring,}
\keyword{network}
\keyword{process}
\keyword{random}
\keyword{statistical}

